zakupki/requirements.txt
asgiref==3.7.2
Django==4.2.4
python-slugify==8.0.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
beautifulsoup4==4.12.2
django-bootstrap-v5==1.0.11
soupsieve==2.4.1
